# Surface Area Heuristic for Bounding Volume Hierarchies