--- a/clovers-cli/src/json_scene.rs
+++ b/clovers-cli/src/json_scene.rs
@@ -1,4 +1,5 @@
-use clovers::scenes::{self, Scene, SceneFile};
+use clovers::bvh::BvhAlgorithm;
+use clovers::scenes::Scene;
 use std::error::Error;
 use std::fs::File;
 use std::io::Read;
@@ -6,8 +7,11 @@ use std::path::Path;
 
 use tracing::info;
 
+use crate::scenefile::SceneFile;
+
 pub fn initialize<'scene>(
     path: &Path,
+    bvh_algorithm: BvhAlgorithm,
     width: u32,
     height: u32,
 ) -> Result<Scene<'scene>, Box<dyn Error>> {
@@ -17,7 +21,7 @@ pub fn initialize<'scene>(
     info!("Parsing the scene file");
     let scene_file: SceneFile = serde_json::from_str(&contents)?;
     info!("Initializing the scene");
-    let scene: Scene = scenes::initialize(scene_file, width, height);
+    let scene: Scene = SceneFile::initialize(scene_file, bvh_algorithm, width, height);
     info!("Count of nodes in the BVH tree: {}", scene.bvh_root.count());
     Ok(scene)
 }