--- a/clovers-cli/src/lib.rs
+++ b/clovers-cli/src/lib.rs
@@ -2,6 +2,7 @@
 
 use clap::Args;
 
+pub mod bvhdepth;
 pub mod colorize;
 pub mod draw_cpu;
 pub mod json_scene;