--- a/clovers/src/bvh/build.rs
+++ b/clovers/src/bvh/build.rs
@@ -1,0 +1,2 @@
+mod longest_axis_midpoint;
+pub use longest_axis_midpoint::build as longest_axis_midpoint;