Implementing the Surface Area Heuristic algorithm for Bounding Volume Hierarchy tree construction.

Based on these two blog articles:
- https://jacco.ompf2.com/2022/04/18/how-to-build-a-bvh-part-2-faster-rays/
- https://jacco.ompf2.com/2022/04/21/how-to-build-a-bvh-part-3-quick-builds/

The new SAH algorithm results in much more efficient BVH trees. This affects especially scenes with a larger amount of primitives.

```
hyperfine -L bvh lam,sah "just cli render -w 1920 -h 1080 --sampler blue --samples 8 --bvh {bvh} -o temp.png ~/sponza.json"
Benchmark 1: just cli render -w 1920 -h 1080 --sampler blue --samples 8 --bvh lam -o temp.png ~/sponza.json
  Time (mean ± σ):     52.065 s ±  0.318 s    [User: 387.275 s, System: 0.751 s]
  Range (min … max):   51.794 s … 52.789 s    10 runs
 
Benchmark 2: just cli render -w 1920 -h 1080 --sampler blue --samples 8 --bvh sah -o temp.png ~/sponza.json
  Time (mean ± σ):     20.147 s ±  0.050 s    [User: 141.036 s, System: 0.381 s]
  Range (min … max):   20.090 s … 20.255 s    10 runs
 
Summary
  just cli render -w 1920 -h 1080 --sampler blue --samples 8 --bvh sah -o temp.png ~/sponza.json ran
    2.58 ± 0.02 times faster than just cli render -w 1920 -h 1080 --sampler blue --samples 8 --bvh lam -o temp.png ~/sponza.json
hyperfine -L bvh lam,sah   5283.53s user 11.72s system 733% cpu 12:02.34 total
```

There's also a couple of new debug rendering modes, for visualizing the BVH node test count and primitive test count needed per pixel.

Before, with Longest Axis Midpoint BVH construction
![1722156865](https://github.com/user-attachments/assets/47b65cea-e562-49c4-8de2-bd59e6710b84)

After, with Surface Area Heuristic BVH construction
![1722156858](https://github.com/user-attachments/assets/2b66eda1-0fd7-4c8f-bcb9-ace84e747045)

Colorization key
<img width="859" alt="Screenshot 2024-07-28 at 12 22 35" src="https://github.com/user-attachments/assets/f62c89df-74bc-4e6d-b4fe-00f4a656f4ed">